New model architecture completed
For each game there's a GameBoard.
A GameBoard have a gameBoardID, a Player, a Grid, a Rack and a ClientProtocol.
A Player have an ID and a name.
A Grid is a matrix of Squares.
A Square have a Tile, a Bonus and a Point coordinate.
A Tile have a Letter, a Value and a coordinate (?)
A Bonus have a value.
A Rack have 7 Tiles.

That's about it. The model is much more complex but way much more efficient.
There's still a lot of work to do in those classes but it's looking good.