Add tile usher and add word method
Not the final version. It just get the letters and the position, format them in a String and send them to the server.
If the server says it's ok, it place the word on the grid.
It doesn't remove the tiles from the rack and doesn't ask for new tiles to the server.
It doesn't check the coordinates.